Formatting changes and moved new detail component to consistent location
